Disable clone.bundle lookup when running repo init
The git repo tool contains an "optimization" where the gerrit server
provides a full tarball of the git repository called a
clone.bundle. repo will use this file when cloning a new repo. Before
bitmap support in git this would have saved a lot of server CPU and
spend up the download. The repo tool attempts to download the
clone.bundle over http and if the server reports 404 it falls back to
regular git clone. When cloning from a Gitlab server however the
download of the clone.bundle returns HTTP 503 code for authentication
redirect and the repo tool exits with error. The repo sync already
disables the clone.bundle fetch.

Signed-off-by: Konrad Scherer <Konrad.Scherer@windriver.com>