setup.py: insert repository "prime" step between "repo init" and "repo sync"
We have linux-yocto and linux-yocto-dev kernel repositories, and they
share objects right up until the v5.10 tag ; the linux-yocto then goes
on with the v5.10.x-stable ; the linux-yocto-dev goes on with mainline
to v5.12 and then v5.12.x-stable currently.

We've already "garbage collected" non-v5.10 and non-v5.12 content out of
each repository, but we still have these sizes for bare clones from the
main git server, as well as the size of a mainline v5.10 repository for
reference:

$du -cshm *
1993    linux-5.10
2058    linux-yocto-dev.git
2027    linux-yocto.git
$du -csh *
2.0G    linux-5.10
2.1G    linux-yocto-dev.git
2.0G    linux-yocto.git

While the absolute sizes are dependent on the server side compression, the
relative ratios won't change, and it doesn't take a rocket scientist to see
that the bulk of the size is the duplicated v5.10 content.

Currently, repo doesn't really do repository sharing at the initial
clone level.  You can use "--mirror" and then "--reference" off of that,
and you can use workspace to have multiple checkouts of the same object
store, but there isn't a way to have a clone repository B start with the
objects of repository A as a baseline starting point.

However, it doesn't throw away object directories if they already exist,
which leaves us a window between "repo init" and "repo sync" where we
can grab a single copy of v5.10 and then "sneak it in" to prime both
repositories.

The use of hardlinks isn't strictly necessary - we'd save the double
download without them, but this way we save space on the customer's
disk as well.  Disk usage in the download after this change is:

$du -csh linux-yocto*
2.2G    linux-yocto-dev.git
80M     linux-yocto.git
2.3G    total

We can see that we are sharing 2G of mainline v5.10 content as the
pack/index files show "2" nlink and the same inode:

$ls -lih linux-yocto*/objects/pack/*
150897461 -r--r--r-- 2 208M linux-yocto-dev.git/objects/pack/pack-1787aeb47c276ad351e1f792245236b2f5f44b8f.idx
150897456 -r--r--r-- 2 1.8G linux-yocto-dev.git/objects/pack/pack-1787aeb47c276ad351e1f792245236b2f5f44b8f.pack
150899141 -r--r--r-- 1 7.3M linux-yocto-dev.git/objects/pack/pack-63f7cec10a893c66daa49cd6064fe3c028b81bc4.idx
150899136 -r--r--r-- 1 213M linux-yocto-dev.git/objects/pack/pack-63f7cec10a893c66daa49cd6064fe3c028b81bc4.pack
150899144 -r--r--r-- 1 3.2M linux-yocto.git/objects/pack/pack-0c019fdfd87280de954f35502a2d2124f4358b26.idx
150899140 -r--r--r-- 1 76M  linux-yocto.git/objects/pack/pack-0c019fdfd87280de954f35502a2d2124f4358b26.pack
150897461 -r--r--r-- 2 208M linux-yocto.git/objects/pack/pack-1787aeb47c276ad351e1f792245236b2f5f44b8f.idx
150897456 -r--r--r-- 2 1.8G linux-yocto.git/objects/pack/pack-1787aeb47c276ad351e1f792245236b2f5f44b8f.pack

and that the unique content for linux-yocto is only 80M in git, and only
220M for the linux-yocto-dev.  (The -dev has the v5.10 --> v5.12 mainline
which makes it the larger of the two).

We take a few minutes for that initial download of the v5.10 objects but
when "repo" comes through to the final lines of the manifest and "clones"
yocto kernels, they populate in seconds, and not downloading the v5.10
content a second time also reduces the total overall download time.

It would be trivial to extend this to other repositories with a large
common ancestry, but for the moment the kernel is the primary focus.

To make this general the default.xml file is parsed and the priming is
only enabled if linux-yocto-dev is present. To handle the release
layout the full path of repository is also retrieved from the
manifest.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Konrad Scherer <Konrad.Scherer@windriver.com>