Add option to disable pruning of git repos during repo sync
--prune has been a default option for repo sync for a long time because it
prevents update problems when setup is called on an existing buildarea. If
an upstream project moves a tag or deletes a branch "foo" and creates
"foo/bar" the repo sync with --prune will fail without a helpful
message. However the --prune option will also delete and locally created
refs, so the new --repo-no-prune flag will allow the user to preserve local
refs at the small risk of the synchronization failing.

Signed-off-by: Konrad Scherer <Konrad.Scherer@windriver.com>