README.sample: remove left over "repo" text
Prior to this commit, we documented using "repo" to update or using
setup.sh to update:

  commit aae0ea9832564520a777412f321e8971caacfc6d
  Author: Mark Hatle <mark.hatle@windriver.com>
  Date:   Mon Oct 24 12:58:07 2016 -0500

    Split Wind River Linux specific content from core functionality

The above commit removed the "repo sync" command but left the old
description of what it did.  The description fragment left on its own
is now just confusing w/o the command context, so lets remove it too.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>