argparse_setup.py: Add --mirror-as-premirrors
Issue: LINUXEXEC-19675

It will be used as:
* --mirror-as-premirrors is not specifed:
  Nothing happened

* Both --mirror --mirror-as-premirrors
  The mirrored project will be *made* as PREMIRRORS

* --mirror-as-premirrors, but no --mirror
  The current project's build will *use* the mirror as PMIRRORS

There are two steps to enable the mirror as PREMIRRORS:
1) Setup the project as mirrored-project:
$ wrlinux-x/setup.sh --dl-layers --all-layers --mirror --mirror-as-premirrors

2) Clone wrlinux-x from mirrored-project/wrlinux-x (local, ssh://, http(s)://
   or git://), and run setup a project2:
$ ./wrlinux-x/setup.sh --all-layers --dl-layers --mirror-as-premirrors

Then the builds based on project2 will only fetch sources from mirrored-project
by default.

Note:
* If project2's wrlinux-x is cloned via "git://" from
mirrored-project/wrlinux-x, then only git based sources can be fetched from the
mirrored-project, this is because git doesn't support download files such as
tarballs, there is a warning for this, so the suggested protocols are other 3
ones.

* The ssh:// protocol may be failed if when BB_NUMBER_THREADS is too large, you
  can set a smaller number in conf/local.conf to fix the issue, e.g.:

BB_NUMBER_THREADS = "4"

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>