askpass/anspass: diable askpass when use ssh
Issue: LIN1022-3665

When using ssh:// protocol to download the product, it's better to
use ssh-agent to manage ssh passphrases. Otherwise, we need to enter
the passphrase many times during setup.

And it's also annoying as the ssh passphrase is in plain text if askpass
enabled because the terminal echoing is diabled only for the text start
with "Password" or end with 'password: ' when askpass enabled.

Signed-off-by: Mingli Yu <mingli.yu@windriver.com>